server/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
hyperliquid-python-sdk
fastapi-limiter
redis
